+++ CHANGES.md
@@ -2,6 +2,7 @@
 =====
 
 - Add `comment`, `pdf2pdf` and `replace` examples.
+- More complete parsing of pandoc.
 
 0.1.0 (2021-04-28)
 =====