+++ CHANGES.md
@@ -1,9 +1,10 @@
-0.1.1 (unreleased)
+0.2.0 (unreleased)
 =====
 
 - Add `comment`, `pdf2pdf` and `replace` examples.
 - More complete parsing of pandoc.
 - Add support for `Div` (#2).
+- Much more complete AST support (#3).
 
 0.1.0 (2021-04-28)
 =====